+++ contracts/protocols/lending/aave/Aave3PoolAdapterBase.sol
@@ -135,7 +135,7 @@ abstract contract Aave3PoolAdapterBase is IPoolAdapter, IPoolAdapterInitializer 
 
     {
       (,,,,, uint256 healthFactor) = _pool.getUserAccountData(address(this));
-      require(healthFactor > controller.MIN_HEALTH_FACTOR2()*10**(18-2), AppErrors.WRONG_HEALTH_FACTOR);
+      require(healthFactor > uint(controller.MIN_HEALTH_FACTOR2())*10**(18-2), AppErrors.WRONG_HEALTH_FACTOR);
     }
 
     // ensure that we have received required borrowed amount, send the amount to the receiver
